% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_PRODIGY_results.R
\name{analyze_PRODIGY_results}
\alias{analyze_PRODIGY_results}
\title{analyze_PRODIGY_results}
\usage{
analyze_PRODIGY_results(all_patients_scores)
}
\arguments{
\item{all_patients_scores}{Either a list of influence scores matrices or a single matrix received by the PRODIGY algorithm.}
}
\value{
A list of ranked genes per sample.
}
\description{
This function analyzes PRODIGY's results (influence score matrices) to rank individual driver genes for a group of patients.
}
\examples{
results = analyze_PRODIGY_results(all_patients_scores) #' @references
}
