% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_performances.R
\name{check_performances}
\alias{check_performances}
\title{check_performances}
\usage{
check_performances(ranked_genes_lists, snv_matrix, gold_standard_drivers)
}
\arguments{
\item{ranked_genes_lists}{A named list of ranked drivers for each patient.}

\item{snv_matrix}{A binary matrix with genes in rows and patients on columns. 1=mutation. All genes must be contained in the global PPI network.}

\item{gold_standard_drivers}{A list of known driver genes to be used as gold standard for validation.}
}
\description{
This function calculates mean precision, recall and F1 of the results provided by PRODIGY against a gold standard driver list
}
\examples{
data(gold_standard_drivers)
check_performances(results,snv_matrix,gold_standard_drivers)
}
