% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODIGY.R
\name{get_network_from_gene_set}
\alias{get_network_from_gene_set}
\title{Return a network induced by a given gene set}
\usage{
get_network_from_gene_set(network, gene_set)
}
\arguments{
\item{network}{A data.table with at least 2 columns. It is assumed that the first two columns contains the genes participating in the interaction}

\item{gene_set}{A vector of gene names.
return A data.table representing the network}
}
\description{
Return a network induced by a given gene set
}
