% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODIGY.R
\name{get_pathway_genes_from_network}
\alias{get_pathway_genes_from_network}
\title{' Return vector of gene names participate in the pathway
' @param pathway A 2 column data.table representing interactions
' return A vector of gene names}
\usage{
get_pathway_genes_from_network(pathway)
}
\description{
' Return vector of gene names participate in the pathway
' @param pathway A 2 column data.table representing interactions
' return A vector of gene names
}
