% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODIGY.R
\name{get_pathway_list_from_graphite}
\alias{get_pathway_list_from_graphite}
\title{Get list of data.table representing pathways using the graphite package. This could take a while so better to use more than a single core}
\usage{
get_pathway_list_from_graphite(
  source = "reactome",
  minimal_number_of_nodes = 10,
  num_of_cores = 1
)
}
\arguments{
\item{source}{A string representing the graphite source to use. Default is Reactome}

\item{minimal_number_of_nodes}{An int, lower bound on the number of pathway genes
return A list of data.tables, each representing a single network}
}
\description{
Get list of data.table representing pathways using the graphite package. This could take a while so better to use more than a single core
}
