% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODIGY_cohort.R
\name{PRODIGY_cohort}
\alias{PRODIGY_cohort}
\title{PRODIGY_cohort}
\usage{
PRODIGY_cohort(snp_matrix, expression_matrix, network = NULL,
  samples = NULL, DEGs = NULL, alpha = 0.05,
  pathwayDB = "reactome", num_of_cores = 1, sample_origins = NULL,
  write_results = F, results_folder = "./")
}
\arguments{
\item{snp_matrix}{A binary matrix with genes in rows and patients on columns. 1=mutation. All genes must be contained in the global PPI network.}

\item{expression_matrix}{A read count matrix with genes in rows and patients on columns. All genes must be contained in the global PPI network.}

\item{network}{The global PPI network. Columns describe the source protein, destination protein and interaction score respectively. The network is considered as undirected.}

\item{DEGs}{Named list of differentially expressed genes for every sample. All genes must be contained in the global PPI network.}

\item{alpha}{The penalty exponent.}

\item{pathwayDB}{The pathway DB name from which curated pathways are taken. Could be one of three built in reservoirs ("reactome","kegg","nci").}

\item{num_of_cores}{The number of CPU cores to be used by the influence scores calculation step.}

\item{sample_origins}{A vector that contains two optional values ("tumor","normal") corresponds to the tissues from which each column in expression_matrix was derived. This vector is utilized for differential expression analysis. If no vector is specified, the sample names of expression_matrix are assumed to be in TCGA format where last two digits correspond to sample type: "01"= solid tumor and "11"= normal.}

\item{write_results}{Should the results be written to text files?}

\item{results_folder}{Location for resulting influence matrices storage (if write_results = T)}

\item{sample}{The sample labels as appears in the SNP and expression matrices.}
}
\value{
A list of influence scores matrices.
}
\description{
This is a wrapper that runs PRODIGY for a cohort of patients.
}
\examples{
data(COAD_SNP)
data(COAD_Expression)
# Load STRING network data 
data(STRING_network)
network = STRING_network
# Take samples for which SNP and expression is available 
samples = intersect(colnames(expression_matrix),colnames(snp_matrix))[1:5]
# Get differentially expressed genes (DEGs) for all samples
expression_matrix = expression_matrix[which(rownames(expression_matrix) \%in\% unique(c(network[,1],network[,2]))),]
DEGs = get_DEGs(expression_matrix,samples,sample_origins=NULL)
# Identify sample origins (tumor or normal)
sample_origins = rep("tumor",ncol(expression_matrix))
sample_origins[substr(colnames(expression_matrix),nchar(colnames(expression_matrix)[1])-1,nchar(colnames(expression_matrix)[1]))=="11"] = "normal"	
# Run PRODIGY
all_patients_scores = PRODIGY_cohort(snp_matrix,expression_matrix,network=network,samples=samples,DEGs=DEGs,alpha=0.05,pathwayDB="reactome",num_of_cores=1,sample_origins=sample_origins)
}
\references{
Love, M. I., Huber, W. & Anders, S. Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2. Genome Biol. 15, 1-21 (2014).
Gabriele Sales, Enrica Calura and Chiara Romualdi, graphite: GRAPH Interaction from pathway Topological Environment (2017).
}
