% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODIGY.R
\name{PRODIGY}
\alias{PRODIGY}
\title{PRODIGY}
\usage{
PRODIGY(snp_matrix, expression_matrix, network = NULL, sample,
  diff_genes = NULL, alpha = 0.05, pathwayDB = "reactome",
  num_of_cores = 1, sample_origins = NULL)
}
\arguments{
\item{snp_matrix}{A binary matrix with genes in rows and patients on columns. 1=mutation. All genes must be contained in the global PPI network.}

\item{expression_matrix}{A read count matrix with genes in rows and patients on columns. All genes must be contained in the global PPI network.}

\item{network}{The global PPI network. Columns describe the source protein, destination protein and interaction score respectively. The network is considered as undirected.}

\item{sample}{The sample label as appears in the SNP and expression matrices.}

\item{diff_genes}{A vector of the sample's differentially expressed genes (with gene names). All genes must be contained in the global PPI network.}

\item{alpha}{the penalty exponent.}

\item{pathwayDB}{The pathway DB name from which curated pathways are taken. Could be one of three built in reservoirs ("reactome","kegg","nci").}

\item{num_of_cores}{The number of CPU cores to be used by the influence scores calculation step.}

\item{sample_origins}{A vector that contains two optional values ("tumor","normal") corresponds to the tissues from which each column in expression_matrix was derived. This vector is utilized for differential expression analysis. If no vector is specified, the sample names of expression_matrix are assumed to be in TCGA format where last two digits correspond to sample type: "01"= solid tumor and "11"= normal.}
}
\value{
A matrix of influence scores for every mutation and every enriched pathway.
}
\description{
This function runs the PRODIGY algorithm for a single patient.
}
\examples{
# Load SNP+expression data from TCGA
load("data/COAD_SNP.RData")
load("data/COAD_Expression.RData")
# Load STRING network data 
load("data/STRING_network.RData")
network = STRING_network
sample = intersect(colnames(expression_matrix),colnames(snp_matrix))[1]
sample_origins = rep("tumor",ncol(expression_matrix))
sample_origins[substr(colnames(expression_matrix),nchar(colnames(expression_matrix)[1])-1,nchar(colnames(expression_matrix)[1]))=="11"] = "normal"	
res = PRODIGY<-function(snp_matrix,expression_matrix,network=network,sample,diff_genes=NULL,alpha=0.05,pathwayDB="reactome",num_of_cores=1,sample_origins = sample_origins)
}
\references{
Love, M. I., Huber, W. & Anders, S. Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2. Genome Biol. 15, 1–21 (2014).
Gabriele Sales, Enrica Calura and Chiara Romualdi, graphite: GRAPH Interaction from pathway Topological Environment (2017).
Gillespie, M., Vastrik, I., Eustachio, P. D., Schmidt, E. & Bono, B. De. Reactome?: a knowledgebase of biological pathways. Nucleic Acids Res. 33, 428–432 (2005).
Schaefer, C. F. et al. PID: The pathway interaction database. Nucleic Acids Res. 37, 674–679 (2009).
Ogata, H. et al. KEGG: Kyoto encyclopedia of genes and genomes. Nucleic Acids Res. 27, 29–34 (1999).
}
